requirements.txt
fastapi
uvicorn
passlib
python-jose
mysql-connector-python